573... c# done... need to verify c++
git-svn-id: http://tradelink.googlecode.com/svn/trunk@2009 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ Responses/BarRequest.cs
@@ -33,7 +33,7 @@ namespace Responses
 
         }
 
-        public override void GotMessage(MessageTypes type, uint source, uint dest, uint msgid, string request, ref string response)
+        public override void GotMessage(MessageTypes type, long source, long dest, long msgid, string request, ref string response)
         {
             if (type== MessageTypes.BARRESPONSE)
                 D(response);
